requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2019.3
requests==2.22.0
six==1.13.0
twilio==6.35.1
urllib3==1.25.7
virtualenv==16.7.9
